Unicode fixorama.
Seriously though, the default encoding for all of serpextract is now utf-8.
Anything passed to the public API functions get_parser, is_serp or extract are
first converted to unicode (with an assumed utf-8 encoding).  We also ensure
that we only work with unicode for the remainder of the function (which has
required the addition of _unicode_parse_qs) and return keywords as unicode.

It's up to the client to then encode into whatever encoding you want.

Eventually, we should probably return keywords in the encoding specified by the
parser which is something we don't use at all right now.